+++ CHANGELOG.md
@@ -3,11 +3,12 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 This change log adheres to standards from [Keep a CHANGELOG](http://keepachangelog.com).
 
-## [Unreleased]
+## [2.0.0] - 2016-07-14
 ### Removed
-- Support for Node < v4
+- **Breaking**: Removed support for Node < v4
 ### Added
 - Added [`no-proxy`] rule
+- Added [`no-mutating-methods`] rule
 ### Fixed
 - Fixed error in [`no-mutating-assign`] report message